Build Linux and Mac wheels on GitHub Actions
This utilizes [cibuildwheel](https://github.com/joerick/cibuildwheel) and GitHub Actions to build Linux and Mac wheels
for CPython 3.5 - 3.8, 64-bit.

The output wheels are uploaded as an artifact which can be downloaded and tested.

Windows will likely be needed at some point as well, and can be added.

Additionally, we can also get GitHub Actions to automatically upload the
wheels to PyPI.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>