Upload wheels to PyPI for tags starting with "v"
This should automatically upload wheels to PyPI when a tag is pushed
that starts with "v". For instance, "v0.0.2".

Note that the version number in the setup.py must be modified as well,
or PyPI will reject the wheels (because they will have the same name
as previous wheels).

We must also create a GitHub secret, "hydroframebot_pypi_token", that
contains a token that will be used to upload to PyPI, before this can
be merged.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>